fix: change service catalogs file name on camelCase, export service catalogs from clients, refactor clients/index.js file with prettier  [WTEL-5667] (https://webitel.atlassian.net/browse/WTEL-5667)
