+++ packages/ui-datalist/src/filters/components/filter-options/service-case/config.js
@@ -1,4 +1,4 @@
-import ServiceCatalogsAPI from '@webitel/ui-sdk/api/clients/caseServiceCatalogs/service-catalogs';
+import ServiceCatalogsAPI from '@webitel/ui-sdk/api/clients/caseServiceCatalogs/serviceCatalogs';
 import ServicesAPI from '@webitel/ui-sdk/api/clients/caseServices/services';
 
 export const searchMethod = ServiceCatalogsAPI.getList;
